move new indexheader cache alongside indexcache for ease of initialization
