+++ pkg/storage/tsdb/indexcache/indexheader/cache.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package index
+package indexheader
 
 import (
 	"context"
