+++ pkg/storage/tsdb/indexcache/indexheader/cache_codec.go
@@ -1,4 +1,4 @@
-package index
+package indexheader
 
 import (
 	"encoding/binary"
