+++ pkg/storage/tsdb/indexcache/indexheader/inmemory_cache_test.go
@@ -1,4 +1,4 @@
-package index
+package indexheader
 
 import (
 	"context"
