+++ pkg/storage/tsdb/indexcache/indexheader/remote_cache.go
@@ -1,4 +1,4 @@
-package index
+package indexheader
 
 import (
 	"context"
